+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/com/azure/messaging/eventhubs/PartitionPumpManager.java
@@ -98,7 +98,7 @@ class PartitionPumpManager {
         List<String> partitionIds = new ArrayList<>(partitionPumps.keySet());
         return Flux.fromIterable(partitionIds)
             .flatMap(partitionId -> {
-                final PartitionPump pump = partitionPumps.get(partitionId);
+                final PartitionPump pump = partitionPumps.remove(partitionId);
                 return pump.closeAsync()
                     .doOnError(ex -> LOGGER.atWarning()
                         .addKeyValue(PARTITION_ID_KEY, partitionId)
@@ -107,8 +107,7 @@ class PartitionPumpManager {
                     .doOnCancel(() -> LOGGER.atWarning()
                         .addKeyValue(PARTITION_ID_KEY, partitionId)
                         .addKeyValue(SIGNAL_TYPE_KEY, SignalType.CANCEL)
-                        .log(Messages.FAILED_CLOSE_CONSUMER_PARTITION))
-                    .doFinally(__ -> partitionPumps.remove(partitionId));
+                        .log(Messages.FAILED_CLOSE_CONSUMER_PARTITION));
             }).then();
     }