+++ sdk/eventhubs/azure-messaging-eventhubs/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 
 - Integrated RequestResponseChannelCache (CBS, Management channel cache) and ReactorSessionCache, these caches are activated when V2 stack is opted-in using the configuration `com.azure.messaging.eventhubs.v2`. ([39107](https://github.com/Azure/azure-sdk-for-java/pull/39107)), ([41805](https://github.com/Azure/azure-sdk-for-java/pull/41805))
+- Add `EventProcessorClient.stop(Duration timeout)` to stop the `EventProcessorClient` and await it to shut down. The `EventProcessorClient.stop()`
+  method now will wait for up to the default timeout (10 seconds) waiting for the processor to stop. ([#41878](https://github.com/Azure/azure-sdk-for-java/pull/41878))
 
 ### Breaking Changes